Backport merge fixes added to ENT
Fix test and address code review comments.
