+++ node/src/integration-test/java/net/corda/serialization/reproduction/GenericReturnFailureReproductionIntegrationTest.java
@@ -2,9 +2,6 @@ package net.corda.serialization.reproduction;
 
 import net.corda.client.rpc.CordaRPCClient;
 import net.corda.core.concurrent.CordaFuture;
-import net.corda.core.flows.FlowLogic;
-import net.corda.core.flows.StartableByRPC;
-import net.corda.core.serialization.CordaSerializable;
 import net.corda.node.services.Permissions;
 import net.corda.testing.driver.Driver;
 import net.corda.testing.driver.DriverParameters;
@@ -14,10 +11,7 @@ import net.corda.testing.node.User;
 import org.junit.Test;
 
 import java.util.Collections;
-import java.util.List;
 import java.util.concurrent.ExecutionException;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class GenericReturnFailureReproductionIntegrationTest {
 
@@ -36,33 +30,7 @@ public class GenericReturnFailureReproductionIntegrationTest {
 
     }
 
-    @StartableByRPC
-    public static class SuperSimpleGenericFlow extends FlowLogic<GenericHolder<String>> {
-        public SuperSimpleGenericFlow() {
-        }
-
-        @Override
-        public GenericHolder<String> call() {
-            return new GenericHolder<>(IntStream.of(100).mapToObj((i) -> "" + i).collect(Collectors.toList()));
-        }
-    }
-
-    @CordaSerializable
-    public static class GenericHolder<S> {
-        private final List<S> items;
-
-        public GenericHolder(List<S> items) {
-            this.items = items;
-        }
-
-        public List<S> getItems() {
-            return items;
-        }
-    }
-
-
     private static <Y> Y getOrThrow(CordaFuture<Y> future) {
-
         try {
             return future.get();
         } catch (InterruptedException | ExecutionException e) {
