+++ tools/shell/src/integration-test/kotlin/net/corda/tools/shell/InteractiveShellIntegrationTest.kt
@@ -148,7 +148,7 @@ class InteractiveShellIntegrationTest {
     @Ignore
     @Test
     fun `ssh runs flows via standalone shell`() {
-        val user = User("u", "p", setOf(Permissions.startFlow<SSHServerTest.FlowICanRun>(),
+        val user = User("u", "p", setOf(Permissions.startFlow<FlowICanRun>(),
                 Permissions.invokeRpc(CordaRPCOps::registeredFlows),
                 Permissions.invokeRpc(CordaRPCOps::nodeInfo)))
         driver(DriverParameters(notarySpecs = emptyList())) {
@@ -191,7 +191,7 @@ class InteractiveShellIntegrationTest {
     @Ignore
     @Test
     fun `ssh run flows via standalone shell over ssl to node`() {
-        val user = User("mark", "dadada", setOf(Permissions.startFlow<SSHServerTest.FlowICanRun>(),
+        val user = User("mark", "dadada", setOf(Permissions.startFlow<FlowICanRun>(),
                 Permissions.invokeRpc(CordaRPCOps::registeredFlows),
                 Permissions.invokeRpc(CordaRPCOps::nodeInfo)/*all()*/))
 
@@ -245,33 +245,6 @@ class InteractiveShellIntegrationTest {
         }
     }
 
-    @Suppress("UNUSED")
-    @StartableByRPC
-    class NoOpFlow : FlowLogic<Unit>() {
-        override val progressTracker = ProgressTracker()
-        override fun call() {
-            println("NO OP!")
-        }
-    }
-
-    @Suppress("UNUSED")
-    @StartableByRPC
-    class NoOpFlowA : FlowLogic<Unit>() {
-        override val progressTracker = ProgressTracker()
-        override fun call() {
-            println("NO OP! (A)")
-        }
-    }
-
-    @Suppress("UNUSED")
-    @StartableByRPC
-    class BurbleFlow : FlowLogic<Unit>() {
-        override val progressTracker = ProgressTracker()
-        override fun call() {
-            println("NO OP! (Burble)")
-        }
-    }
-
     @Test
     fun `shell should start flow with fully qualified class name`() {
         val user = User("u", "p", setOf(all()))
@@ -298,7 +271,7 @@ class InteractiveShellIntegrationTest {
                 on { render(any(), any()) }  doAnswer { InteractiveShell.latch.countDown() }
             }
             InteractiveShell.runFlowByNameFragment(
-                    InteractiveShellIntegrationTest::class.qualifiedName + "\$NoOpFlow",
+                    "NoOpFlow",
                     "", output, node.rpc, ansiProgressRenderer)
         }
         assertThat(successful).isTrue()
@@ -330,7 +303,7 @@ class InteractiveShellIntegrationTest {
                 on { render(any(), any()) }  doAnswer { InteractiveShell.latch.countDown() }
             }
             InteractiveShell.runFlowByNameFragment(
-                    "InteractiveShellIntegrationTest\$NoOpFlowA",
+                    "NoOpFlowA",
                     "", output, node.rpc, ansiProgressRenderer)
         }
         assertThat(successful).isTrue()
@@ -362,7 +335,7 @@ class InteractiveShellIntegrationTest {
                 on { render(any(), any()) }  doAnswer { InteractiveShell.latch.countDown() }
             }
             InteractiveShell.runFlowByNameFragment(
-                    InteractiveShellIntegrationTest::class.qualifiedName + "\$NoOpFlo",
+                    "NoOpFlo",
                     "", output, node.rpc, ansiProgressRenderer)
         }
         assertThat(successful).isTrue()
@@ -398,5 +371,32 @@ class InteractiveShellIntegrationTest {
                     "", output, node.rpc, ansiProgressRenderer)
         }
         assertThat(successful).isTrue()
+    }
+}
+
+@Suppress("UNUSED")
+@StartableByRPC
+class NoOpFlow : FlowLogic<Unit>() {
+    override val progressTracker = ProgressTracker()
+    override fun call() {
+        println("NO OP!")
+    }
+}
+
+@Suppress("UNUSED")
+@StartableByRPC
+class NoOpFlowA : FlowLogic<Unit>() {
+    override val progressTracker = ProgressTracker()
+    override fun call() {
+        println("NO OP! (A)")
+    }
+}
+
+@Suppress("UNUSED")
+@StartableByRPC
+class BurbleFlow : FlowLogic<Unit>() {
+    override val progressTracker = ProgressTracker()
+    override fun call() {
+        println("NO OP! (Burble)")
     }
 }
