+++ tools/shell/src/test/kotlin/net/corda/tools/shell/InteractiveShellTest.kt
@@ -110,20 +110,6 @@ class InteractiveShellTest {
                 """.trimIndent()
     }
 
-    @Suppress("UNUSED")
-    class FlowA(val a: String) : FlowLogic<String>() {
-        constructor(b: Int?) : this(b.toString())
-        constructor(b: Int?, c: String) : this(b.toString() + c)
-        constructor(amount: Amount<Currency>) : this(amount.toString())
-        constructor(pair: Pair<Amount<Currency>, SecureHash.SHA256>) : this(pair.toString())
-        constructor(party: Party) : this(party.name.toString())
-        constructor(b: Int?, amount: Amount<UserValue>) : this("${(b ?: 0) + amount.quantity} ${amount.token}")
-        constructor(b: Array<String>) : this(b.joinToString("+"))
-        constructor(amounts: Array<Amount<UserValue>>) : this(amounts.joinToString("++", transform = Amount<UserValue>::toString))
-
-        override val progressTracker = ProgressTracker()
-        override fun call() = a
-    }
 
     private val ids = InMemoryIdentityService(listOf(megaCorp.identity), DEV_ROOT_CA.certificate)
     @Suppress("DEPRECATION")
@@ -148,7 +134,7 @@ class InteractiveShellTest {
 
         return objectMapper
     }
-    
+
     @Test
     fun flowStartSimple() {
         check("a: Hi there", "Hi there")
@@ -161,26 +147,26 @@ class InteractiveShellTest {
 
     @Test
     fun flowStartWithNestedTypes() = check(
-        input = "pair: { first: $100.12, second: df489807f81c8c8829e509e1bcb92e6692b9dd9d624b7456435cb2f51dc82587 }",
-        expected = "(100.12 USD, DF489807F81C8C8829E509E1BCB92E6692B9DD9D624B7456435CB2F51DC82587)"
+            input = "pair: { first: $100.12, second: df489807f81c8c8829e509e1bcb92e6692b9dd9d624b7456435cb2f51dc82587 }",
+            expected = "(100.12 USD, DF489807F81C8C8829E509E1BCB92E6692B9DD9D624B7456435CB2F51DC82587)"
     )
 
     @Test
     fun flowStartWithArrayType() = check(
-        input = "b: [ One, Two, Three, Four ]",
-        expected = "One+Two+Three+Four"
+            input = "b: [ One, Two, Three, Four ]",
+            expected = "One+Two+Three+Four"
     )
 
     @Test
     fun flowStartWithUserAmount() = check(
-        input = """b: 500, amount: { "quantity": 10001, "token":{ "label": "of value" } }""",
-        expected = "10501 of value"
+            input = """b: 500, amount: { "quantity": 10001, "token":{ "label": "of value" } }""",
+            expected = "10501 of value"
     )
 
     @Test
     fun flowStartWithArrayOfNestedTypes() = check(
-        input = """amounts: [ { "quantity": 10, "token": { "label": "(1)" } }, { "quantity": 200, "token": { "label": "(2)" } } ]""",
-        expected = "10 (1)++200 (2)"
+            input = """amounts: [ { "quantity": 10, "token": { "label": "(1)" } }, { "quantity": 200, "token": { "label": "(2)" } } ]""",
+            expected = "10 (1)++200 (2)"
     )
 
     @Test(expected = InteractiveShell.NoApplicableConstructor::class)
@@ -198,11 +184,11 @@ class InteractiveShellTest {
             check("", expected = "")
         }
         val correct = setOf(
-                "[amounts: Amount<InteractiveShellTest.UserValue>[]]: missing parameter amounts",
+                "[amounts: Amount<UserValue>[]]: missing parameter amounts",
                 "[amount: Amount<Currency>]: missing parameter amount",
                 "[pair: Pair<Amount<Currency>, SecureHash.SHA256>]: missing parameter pair",
                 "[party: Party]: missing parameter party",
-                "[b: Integer, amount: Amount<InteractiveShellTest.UserValue>]: missing parameter b",
+                "[b: Integer, amount: Amount<UserValue>]: missing parameter b",
                 "[b: String[]]: missing parameter b",
                 "[b: Integer, c: String]: missing parameter b",
                 "[a: String]: missing parameter a",
@@ -245,8 +231,24 @@ class InteractiveShellTest {
         verify(printWriter).println(NETWORK_MAP_JSON_PAYLOAD.replace("\n", System.lineSeparator()))
     }
 
-    @ToStringSerialize
-    data class UserValue(@JsonProperty("label") val label: String) {
-        override fun toString() = label
-    }
+}
+
+@ToStringSerialize
+data class UserValue(@JsonProperty("label") val label: String) {
+    override fun toString() = label
+}
+
+@Suppress("UNUSED")
+class FlowA(val a: String) : FlowLogic<String>() {
+    constructor(b: Int?) : this(b.toString())
+    constructor(b: Int?, c: String) : this(b.toString() + c)
+    constructor(amount: Amount<Currency>) : this(amount.toString())
+    constructor(pair: Pair<Amount<Currency>, SecureHash.SHA256>) : this(pair.toString())
+    constructor(party: Party) : this(party.name.toString())
+    constructor(b: Int?, amount: Amount<UserValue>) : this("${(b ?: 0) + amount.quantity} ${amount.token}")
+    constructor(b: Array<String>) : this(b.joinToString("+"))
+    constructor(amounts: Array<Amount<UserValue>>) : this(amounts.joinToString("++", transform = Amount<UserValue>::toString))
+
+    override val progressTracker = ProgressTracker()
+    override fun call() = a
 }
